+++ internal/httpserver/handlers_challenge.go
@@ -1,0 +1,6 @@
+package httpserver
+
+import "net/http"
+
+func (h *handlers) issueChallenge(w http.ResponseWriter, r *http.Request)  { notImplemented(w, r) }
+func (h *handlers) verifyChallenge(w http.ResponseWriter, r *http.Request) { notImplemented(w, r) }